+++ ts-simple-ast-extra/src/diagnostics.ts
@@ -16,3 +16,30 @@ function print(s: DiagnosticMessageChain | undefined): string {
     return `${s.getMessageText()} - ${print(s.getNext())}`
   }
 }
+
+// export interface GetDiagnosticsOptions {
+//   type: 'syntactical'|'semantical'
+//   project:Project
+// }
+// export function getDiagnostics(o: GetDiagnosticsOptions){
+//   return o.type==='semantical' ? o.project.getProgram().getSemanticDiagnostics() : 
+// }
+
+
+    // /**
+    //  * Gets the pre-emit diagnostics.
+    //  * @param sourceFile - Optional source file to filter the results by.
+    //  */
+    // getSyntacticDiagnostics(sourceFile?: SourceFile): Diagnostic[] {
+    //     const compilerDiagnostics = this.program.compilerObject.getSyntacticDiagnostics(sourceFile == null ? undefined : sourceFile.compilerNode);
+    //     return compilerDiagnostics.map(d => this.compilerFactory.getDiagnostic(d));
+    // }
+
+    // /**
+    //  * Gets the pre-emit diagnostics.
+    //  * @param sourceFile - Optional source file to filter the results by.
+    //  */
+    // getSemanticDiagnostics(sourceFile?: SourceFile): Diagnostic[] {
+    //     const compilerDiagnostics = this.program.compilerObject.getSemanticDiagnostics(sourceFile == null ? undefined : sourceFile.compilerNode);
+    //     return compilerDiagnostics.map(d => this.compilerFactory.getDiagnostic(d));
+    // }