+++ ts-simple-ast-extra/src/refactor/format.ts
@@ -1,28 +1,40 @@
-import { RemoveProperties } from 'misc-utils-of-mine-generic'
+import { RemoveProperties, checkThrow } from 'misc-utils-of-mine-generic'
 import { Project, SourceFile } from 'ts-morph'
-import { FormatCodeSettings, getDefaultFormatCodeSettings } from 'typescript'
+import { FormatCodeSettings, getDefaultFormatCodeSettings, DiagnosticCategory } from 'typescript'
 import { organizeImports, OrganizeImportsOptions } from './organizeImports'
 import { quotes, QuotesOptions } from './quotes'
 import { trailingSemicolons, TrailingSemicolonsOptions } from './trailingSemicolons'
+import { EmptyLinesOptions, emptyLines } from './emptyLines';
+import { getDiagnosticMessage, printDiagnostics } from '../diagnostics';
 
-export interface RefactorFormatBaseOptions extends Partial<FormatCodeSettings> {
+export interface RefactorInputOptions {
   file: SourceFile
   project: Project
 }
 
-export interface FormatOptions extends RefactorFormatBaseOptions, QuotesOptions, TrailingSemicolonsOptions,
-  RemoveProperties<OrganizeImportsOptions, 'quotePreference'> {
+export interface RefactorBaseOptions {
+  verifyErrors?: 'all' | 'syntactical' | 'semantical'
+}
+
+export interface RefactorFormatBaseOptions extends Partial<FormatCodeSettings>, RefactorInputOptions {
+
+}
+
+export interface FormatOptions extends RefactorBaseOptions, RefactorFormatBaseOptions, QuotesOptions, TrailingSemicolonsOptions,
+  RemoveProperties<OrganizeImportsOptions, 'quotePreference'>, EmptyLinesOptions {
 
 }
 
 export function format(options: FormatOptions) {
   options = Object.assign({}, getDefaultFormatCodeSettings(options.newLineCharacter || '\n'), options)
-
+  if (options.verifyErrors) {
+    const d = options.verifyErrors==='all' ? options.project.getPreEmitDiagnostics() : options.verifyErrors==='semantical' ?  options.project.getProgram().getSemanticDiagnostics() : options.verifyErrors==='syntactical' ?  options.project.getProgram().getSyntacticDiagnostics() : [] 
+    checkThrow(d.length === 0, `TypeScript errors found and verifyErrors === '${options.verifyErrors}' was used. Aborting. \nErrors:\n * ${d.map(getDiagnosticMessage).join('\n * ')}`)
+  }
   let file = organizeImports(options) // Important: this first since TS won't respect formatSettings and do "heuristics"
-  options = { ...options, file: file || options.file } // regenerate options since it forgets ts-morph nodes.
   trailingSemicolons(options)
   quotes(options)
-
+  emptyLines(options)
   const edits = options.project
     .getLanguageService()
     .getFormattingEditsForDocument(options.file.getSourceFile().getFilePath(), options)
@@ -30,7 +42,7 @@ export function format(options: FormatOptions) {
   return file
 }
 
-interface FormatBaseNoInput extends RemoveProperties<FormatOptions, 'file' | 'project'> {
+interface FormatBaseNoInput extends RemoveProperties<FormatOptions, keyof RefactorInputOptions> {
 
 }